move dashboard to /dashboard route
